custom_images/ansible_bootstrapper/requirements.txt
ansible>=8.4.0
docker-py>=1.10.0
pyyaml